Outbox: Add testcase with common
