+++ tests/bootstrap.php
@@ -98,6 +98,7 @@ function http_disable_request( $response, $args, $url ) {
 
 // Start up the WP testing environment.
 require $_tests_dir . '/includes/bootstrap.php';
+require __DIR__ . '/class-activitypub-outbox-testcase.php';
 require __DIR__ . '/class-activitypub-testcase-cache-http.php';
 require __DIR__ . '/class-test-rest-controller-testcase.php';
